Refactor player.rb to split player subclasses into seaparate files, along with corresponding separate files for previous tests.
