+++ lib/human_player.rb
@@ -1,0 +1,16 @@
+require_relative './player.rb'
+
+class HumanPlayer < Player
+
+  attr_accessor :marker, :name 
+
+  def provide_name
+    @name = gets.chomp
+  end
+ 
+  def make_selection 
+    user_selection = gets.chomp
+    user_selection
+  end
+  
+end
