+++ lib/player.rb
@@ -8,27 +8,3 @@ class Player
   end
   
 end
-
-class HumanPlayer < Player
-
-  attr_accessor :marker, :name # why does subclass not inherit attr from superclass?
-
-  def provide_name
-    @name = gets.chomp
-  end
- 
-  def make_selection 
-    user_selection = gets.chomp
-    user_selection
-  end
-  
-end
-
-class ThreeByThreeTTTComputerPlayer < Player
-
-  def select_box(board)
-    index_selected = board.find_index { |box| box == " " }
-    index_selected + 1
-  end
-
-end
