+++ lib/three_by_three_ttt_computer_player.rb
@@ -1,0 +1,10 @@
+require_relative './player.rb'
+
+class ThreeByThreeTTTComputerPlayer < Player
+
+  def select_box(board)
+    index_selected = board.find_index { |box| box == " " }
+    index_selected + 1
+  end
+
+end
