+++ spec/player_spec.rb
@@ -39,52 +39,23 @@ describe Player do
   end
 end
 
-describe HumanPlayer do
-
-  it "is a subclass of the Player class" do
-    expect(HumanPlayer.superclass).to eq(Player)
-  end
-
-  describe "#provide_name" do
-    
-    it "gets the input of the user for its name and sets the @name instance variable" do
-      @player = HumanPlayer.new
-      allow(@player).to receive(:gets).and_return("Betty")
-      @player.provide_name
-      expect(@player.name).to eq("Betty")
-    end
-  
-  end
-
-  describe "#make_selection" do 
-
-    it "gets the input of the user and returns it" do 
-      @player = HumanPlayer.new
-      allow(@player).to receive(:gets).and_return("X")
-      @player.make_selection
-      expect(@player.make_selection).to eq("X")
-    end
-
-  end
-end
-
-describe ThreeByThreeTTTComputerPlayer do
-
-  let(:computer_player) { ThreeByThreeTTTComputerPlayer.new }
-
-  it "is a subclass of the Player class" do
-    expect(ThreeByThreeTTTComputerPlayer.superclass).to eq(Player)
-  end
-
-  describe "#select_box(board)" do
-    it "selects the first box of the board when passed a board array with all empty boxes" do
-      board = [" ", " ", " ", " ", " ", " ", " ", " ", " "]
-      expect(computer_player.select_box(board)).to eq(1)
-    end
-
-    it "selects the first empty box of the board when passed a board array with some moves made already" do 
-      board = ["X", "O", " ", " ", " ", " ", " ", " ", " "]
-      expect(computer_player.select_box(board)).to eq(3)
-    end
-  end
-end
+#describe ThreeByThreeTTTComputerPlayer do
+#
+#  let(:computer_player) { ThreeByThreeTTTComputerPlayer.new }
+#
+#  it "is a subclass of the Player class" do
+#    expect(ThreeByThreeTTTComputerPlayer.superclass).to eq(Player)
+#  end
+#
+#  describe "#select_box(board)" do
+#    it "selects the first box of the board when passed a board array with all empty boxes" do
+#      board = [" ", " ", " ", " ", " ", " ", " ", " ", " "]
+#      expect(computer_player.select_box(board)).to eq(1)
+#    end
+#
+#    it "selects the first empty box of the board when passed a board array with some moves made already" do 
+#      board = ["X", "O", " ", " ", " ", " ", " ", " ", " "]
+#      expect(computer_player.select_box(board)).to eq(3)
+#    end
+#  end
+# end
